# specs/002-fullstack-web-app/contracts/api.yaml
openapi: 3.1.0
info:
  title: Todo API
  description: RESTful API for the Full-Stack Web Todo Application
  version: 1.0.0
  contact:
    name: Evolution of Todo

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todo.example.com
    description: Production server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        user_id:
          type: string
          description: Owner's user ID
          example: "cuid_abc123"
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-12-02T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-12-02T10:00:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Completion status
          example: true

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Missing authentication"

    Forbidden:
      description: User not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authorized to access this task"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string

paths:
  /api/tasks:
    get:
      summary: List tasks
      description: Get all tasks for the authenticated user
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by completion status
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of tasks to skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 100
          description: Maximum number of tasks to return
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create task
      description: Create a new task for the authenticated user
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    get:
      summary: Get task
      description: Get a specific task by ID
      operationId: getTask
      tags:
        - Tasks
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update task
      description: Update a task's title, description, or completion status
      operationId: updateTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete task
      description: Delete a task permanently
      operationId: deleteTask
      tags:
        - Tasks
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/tasks/{task_id}/complete:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    patch:
      summary: Toggle task completion
      description: Toggle the completion status of a task
      operationId: toggleTaskComplete
      tags:
        - Tasks
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/health:
    get:
      summary: Health check
      description: Check if the API is running
      operationId: healthCheck
      tags:
        - System
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"